extract trace formatting helpers
