+++ src/lowering/emit.ts
@@ -83,6 +83,15 @@ import { sizeOfTypeExpr } from '../semantics/layout.js';
 import { encodeInstruction } from '../z80/encode.js';
 import type { OpStackPolicyMode } from '../pipeline.js';
 import { loadBinInput, loadHexInput } from './inputAssets.js';
+import {
+  formatAbs16FixupAsm,
+  formatAbs16FixupEdAsm,
+  formatAbs16FixupPrefixedAsm,
+  formatAsmInstrForTrace,
+  formatIxDisp,
+  toHexByte,
+  toHexWord,
+} from './traceFormat.js';
 
 function diag(diagnostics: Diagnostic[], file: string, message: string): void {
   diagnostics.push({ id: DiagnosticIds.EmitError, severity: 'error', message, file });
@@ -393,222 +402,6 @@ export function emitProgram(
     codeSourceSegments.push({ ...currentCodeSegmentTag, start, end });
   };
 
-  const toHexByte = (n: number): string =>
-    `$${(n & 0xff).toString(16).toUpperCase().padStart(2, '0')}`;
-  const toHexWord = (n: number): string =>
-    `$${(n & 0xffff).toString(16).toUpperCase().padStart(4, '0')}`;
-
-  const formatImmExprForAsm = (expr: ImmExprNode): string => {
-    switch (expr.kind) {
-      case 'ImmLiteral':
-        return toHexWord(expr.value);
-      case 'ImmName':
-        return expr.name;
-      case 'ImmSizeof':
-        return 'sizeof(...)';
-      case 'ImmOffsetof':
-        return 'offsetof(...)';
-      case 'ImmUnary':
-        return `${expr.op}${formatImmExprForAsm(expr.expr)}`;
-      case 'ImmBinary':
-        return `${formatImmExprForAsm(expr.left)} ${expr.op} ${formatImmExprForAsm(expr.right)}`;
-      default:
-        return 'imm';
-    }
-  };
-
-  const formatEaExprForAsm = (ea: EaExprNode): string => {
-    switch (ea.kind) {
-      case 'EaName':
-        return ea.name;
-      case 'EaField':
-        return `${formatEaExprForAsm(ea.base)}.${ea.field}`;
-      case 'EaAdd':
-        return `${formatEaExprForAsm(ea.base)} + ${formatImmExprForAsm(ea.offset)}`;
-      case 'EaSub':
-        return `${formatEaExprForAsm(ea.base)} - ${formatImmExprForAsm(ea.offset)}`;
-      case 'EaIndex': {
-        let idx = '';
-        switch (ea.index.kind) {
-          case 'IndexImm':
-            idx = formatImmExprForAsm(ea.index.value);
-            break;
-          case 'IndexReg8':
-          case 'IndexReg16':
-            idx = ea.index.reg;
-            break;
-          case 'IndexMemHL':
-            idx = '(HL)';
-            break;
-          case 'IndexMemIxIy':
-            idx = ea.index.disp
-              ? `${ea.index.base}${ea.index.disp.kind === 'ImmUnary' ? '' : '+'}${formatImmExprForAsm(ea.index.disp)}`
-              : ea.index.base;
-            break;
-          case 'IndexEa':
-            idx = formatEaExprForAsm(ea.index.expr);
-            break;
-        }
-        return `${formatEaExprForAsm(ea.base)}[${idx}]`;
-      }
-      default:
-        return 'ea';
-    }
-  };
-
-  const formatAsmOperandForTrace = (operand: AsmOperandNode): string => {
-    switch (operand.kind) {
-      case 'Reg':
-        return operand.name;
-      case 'Imm':
-        return formatImmExprForAsm(operand.expr);
-      case 'Ea':
-        return formatEaExprForAsm(operand.expr);
-      case 'Mem':
-        return `(${formatEaExprForAsm(operand.expr)})`;
-      case 'PortC':
-        return '(C)';
-      case 'PortImm8':
-        return `(${formatImmExprForAsm(operand.expr)})`;
-      default:
-        return '?';
-    }
-  };
-
-  const formatAsmInstrForTrace = (head: string, operands: AsmOperandNode[]): string => {
-    const lowerHead = head.toLowerCase();
-    if (operands.length === 0) return lowerHead;
-    return `${lowerHead} ${operands.map(formatAsmOperandForTrace).join(', ')}`;
-  };
-
-  const formatFixupSymbolExpr = (baseLower: string, addend: number): string => {
-    if (addend === 0) return baseLower;
-    if (addend > 0) return `${baseLower} + ${addend}`;
-    return `${baseLower} - ${Math.abs(addend)}`;
-  };
-
-  const jpCondFromOpcode = (opcode: number): string | undefined => {
-    switch (opcode & 0xff) {
-      case 0xc2:
-        return 'NZ';
-      case 0xca:
-        return 'Z';
-      case 0xd2:
-        return 'NC';
-      case 0xda:
-        return 'C';
-      case 0xe2:
-        return 'PO';
-      case 0xea:
-        return 'PE';
-      case 0xf2:
-        return 'P';
-      case 0xfa:
-        return 'M';
-      default:
-        return undefined;
-    }
-  };
-
-  const callCondFromOpcode = (opcode: number): string | undefined => {
-    switch (opcode & 0xff) {
-      case 0xc4:
-        return 'NZ';
-      case 0xcc:
-        return 'Z';
-      case 0xd4:
-        return 'NC';
-      case 0xdc:
-        return 'C';
-      case 0xe4:
-        return 'PO';
-      case 0xec:
-        return 'PE';
-      case 0xf4:
-        return 'P';
-      case 0xfc:
-        return 'M';
-      default:
-        return undefined;
-    }
-  };
-
-  const formatAbs16FixupAsm = (opcode: number, baseLower: string, addend: number): string => {
-    const sym = formatFixupSymbolExpr(baseLower, addend);
-    switch (opcode & 0xff) {
-      case 0x01:
-        return `ld BC, ${sym}`;
-      case 0x11:
-        return `ld DE, ${sym}`;
-      case 0x21:
-        return `ld HL, ${sym}`;
-      case 0x31:
-        return `ld SP, ${sym}`;
-      case 0x2a:
-        return `ld HL, (${sym})`;
-      case 0x3a:
-        return `ld A, (${sym})`;
-      case 0x22:
-        return `ld (${sym}), HL`;
-      case 0x32:
-        return `ld (${sym}), A`;
-      case 0xc3:
-        return `jp ${sym}`;
-      case 0xcd:
-        return `call ${sym}`;
-      default: {
-        const jpCc = jpCondFromOpcode(opcode);
-        if (jpCc) return `jp ${jpCc}, ${sym}`;
-        const callCc = callCondFromOpcode(opcode);
-        if (callCc) return `call ${callCc}, ${sym}`;
-        return `db ${toHexByte(opcode)}, lo(${baseLower}), hi(${baseLower})`;
-      }
-    }
-  };
-
-  const formatAbs16FixupEdAsm = (opcode2: number, baseLower: string, addend: number): string => {
-    const sym = formatFixupSymbolExpr(baseLower, addend);
-    switch (opcode2 & 0xff) {
-      case 0x4b:
-        return `ld BC, (${sym})`;
-      case 0x5b:
-        return `ld DE, (${sym})`;
-      case 0x7b:
-        return `ld SP, (${sym})`;
-      case 0x43:
-        return `ld (${sym}), BC`;
-      case 0x53:
-        return `ld (${sym}), DE`;
-      case 0x73:
-        return `ld (${sym}), SP`;
-      default:
-        return `db $ED, ${toHexByte(opcode2)}, lo(${baseLower}), hi(${baseLower})`;
-    }
-  };
-
-  const formatAbs16FixupPrefixedAsm = (
-    prefix: number,
-    opcode2: number,
-    baseLower: string,
-    addend: number,
-  ): string => {
-    const sym = formatFixupSymbolExpr(baseLower, addend);
-    const reg16 = prefix === 0xdd ? 'IX' : prefix === 0xfd ? 'IY' : undefined;
-    if (!reg16) {
-      return `db ${toHexByte(prefix)}, ${toHexByte(opcode2)}, lo(${baseLower}), hi(${baseLower})`;
-    }
-    switch (opcode2 & 0xff) {
-      case 0x21:
-        return `ld ${reg16}, ${sym}`;
-      case 0x2a:
-        return `ld ${reg16}, (${sym})`;
-      case 0x22:
-        return `ld (${sym}), ${reg16}`;
-      default:
-        return `db ${toHexByte(prefix)}, ${toHexByte(opcode2)}, lo(${baseLower}), hi(${baseLower})`;
-    }
-  };
-
   const traceInstruction = (offset: number, bytesOut: Uint8Array, text: string): void => {
     if (bytesOut.length === 0) return;
     codeAsmTrace.push({
@@ -810,12 +603,6 @@ export function emitProgram(
       return false;
     }
     return emitInstr('push', [{ kind: 'Reg', span, name: 'HL' }], span);
-  };
-
-  const formatIxDisp = (disp: number): string => {
-    const hex = Math.abs(disp).toString(16).padStart(2, '0');
-    const sign = disp >= 0 ? '+' : '-';
-    return `${sign}$${hex}`;
   };
 
   const emitStepPipeline = (pipe: StepPipeline, span: SourceSpan): boolean => {
